requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
datefinder==0.7.3
defusedxml==0.7.1
Django==4.1.6
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.0
frozenlist==1.3.3
gunicorn==20.0.4
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
oauthlib==3.2.2
openapi-codec==1.3.2
outcome==1.2.0
packaging==23.0
pd==0.0.4
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.8.0
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
wsproto==1.2.0
yarl==1.8.2
